reorganize VectorField
