+++ dynamo/plot/topography.py
@@ -20,11 +20,11 @@ from ..tools.cell_velocities import cell_velocities
 from ..tools.utils import nearest_neighbors, update_dict
 from ..vectorfield.scVectorField import BaseVectorField
 from ..vectorfield.topography import (  # , compute_separatrices
-    VectorField,
     VectorField2D,
 )
 from ..vectorfield.topography import topography as _topology  # , compute_separatrices
 from ..vectorfield.utils import vecfld_from_adata
+from ..vectorfield.VectorField import VectorField
 from ..vectorfield.vector_calculus import curl, divergence
 from .scatters import docstrings, scatters, scatters_interactive
 from .utils import (
